One more fresh Mac project
There are some landmines.  (Once UseXamMacFullFramework is picked once,
there doesn't appear to be a way to turn that off via Visual Studio for
Mac anymore and the executable balloons from 9MB to 35MB!)  So now we
have a baseline.

This solution also lives in the right place now and overwrites the old
false-start from a few years ago.