Finish support for MusicXML files
This closes #6.  The last little bit was the now-finicky .xml file extension.  Historically it has been used as both a top-level Synthesia metadata file extension but could also now mean a MusicXML file using one of that file type's older accepted file extensions.  If you drag in more than one file, anything with .xml is assumed to be a MusicXML song and is added to the list.  If you only drag in a single file, it checks to see which kind it is and either opens it as a metadata file or adds it as a song accordingly.